+++ src/AttributeInfrastructure/DataAttributeInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator;
+namespace Yiisoft\Hydrator\AttributeInfrastructure;
 
 /**
  * An interface for data attributes (allowed only in classes). Can be used to change the way data to be assigned is
